+++ spec/rails_app/app/controllers/sorcery_controller.rb
@@ -141,6 +141,10 @@ class SorceryController < ActionController::Base
     login_at(:facebook, state: 'bla')
   end
 
+  def login_at_test_instagram
+    login_at(:instagram)
+  end
+
   def test_login_from_twitter
     if @user = login_from(:twitter)
       redirect_to 'bla', notice: 'Success!'
@@ -239,6 +243,14 @@ class SorceryController < ActionController::Base
     end
   end
 
+  def test_login_from_instagram
+    if @user = login_from(:instagram)
+      redirect_to 'bla', notice: 'Success!'
+    else
+      redirect_to 'blu', alert: 'Failed!'
+    end
+  end
+
   def test_return_to_with_external_twitter
     if @user = login_from(:twitter)
       redirect_back_or_to 'bla', notice: 'Success!'
@@ -331,6 +343,14 @@ class SorceryController < ActionController::Base
 
   def test_return_to_with_external_slack
     if @user = login_from(:slack)
+      redirect_back_or_to 'bla', notice: 'Success!'
+    else
+      redirect_to 'blu', alert: 'Failed!'
+    end
+  end
+
+  def test_return_to_with_external_instagram
+    if @user = login_from(:instagram)
       redirect_back_or_to 'bla', notice: 'Success!'
     else
       redirect_to 'blu', alert: 'Failed!'